backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
fpdf==1.7.2
